Separated style property & value into shared style type
