+++ components/style/src/lib.rs
@@ -1,12 +1,5 @@
 pub mod expand;
-pub mod inheritable;
-pub mod property;
 pub mod render_tree;
 pub mod selector_matching;
 pub mod tree_builder;
-pub mod value;
 pub mod value_processing;
-pub mod values;
-
-#[macro_use]
-extern crate lazy_static;
