+++ components/style_types/src/value.rs
@@ -1,9 +1,6 @@
 use css::{parser::structs::ComponentValue, tokenizer::token::Token};
 
-use crate::{
-    property::Property,
-    value_processing::{ComputeContext, ValueRef},
-};
+use crate::property::Property;
 
 use super::values::prelude::*;
 
@@ -271,17 +268,6 @@ impl Value {
             // TODO: replace with `medium` when we support absolute size
             Property::FontSize => Value::Length(Length::new_px(16.)),
             Property::TextAlign => Value::TextAlign(TextAlign::Left),
-        }
-    }
-
-    pub fn compute(&self, property: &Property, context: &mut ComputeContext) -> ValueRef {
-        match self {
-            Value::Color(color) => color.compute(property, context),
-            Value::BorderRadius(border_radius) => border_radius.compute(property, context),
-            Value::BorderWidth(border_width) => border_width.compute(property, context),
-            Value::Length(length) => length.compute(property, context),
-            Value::Percentage(percentage) => percentage.compute(property, context),
-            _ => context.style_cache.get(&self),
         }
     }
 }
