+++ components/style_types/src/values/text_align.rs
